requirements.txt
# Core ML
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
scikit-learn==1.4.2

# Experiment Tracking
mlflow==2.12.1

# API / Inference
fastapi==0.110.2
uvicorn==0.29.0
pillow==10.3.0
python-multipart==0.0.9

# Monitoring
prometheus-client==0.20.0

# Testing
# pytest==8.2.0

# Visualization (for notebook only)
matplotlib==3.8.4
seaborn==0.13.2

# Optional but recommended
# pandas==2.2.2
